requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==5.1
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
gunicorn==20.0.4
idna==3.7
itypes==1.2.0
Jinja2==3.1.4
Markdown==3.7
MarkupSafe==2.1.5
numpy==1.26.4
opencv-python==4.7.0.72
pathlib==1.0.1
pillow==10.4.0
Pygments==2.7.4
PyJWT==1.7.1
pytz==2024.1
requests==2.32.3
setuptools==69.2.0
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
wheel==0.43.0
whitenoise==5.0.1
